refactor: 데이터 레이어를 data/ 디렉토리로 분리
- components/*/data.ts(x) → data/*.ts(x)로 이동
- 컴포넌트 import 경로를 @/data/로 변경
- 타입 정의는 각 data 파일에 co-locate 유지
