+++ data/educations.tsx
